requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
decorator==5.1.1
distrax==0.1.5
dm-haiku==0.0.11
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.6.0
flax==0.7.5
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-resources==6.1.1
jax==0.4.23
jaxlib==0.4.23
jmp==0.0.4
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.3.2
ml_collections==0.1.1
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.26.3
omegaconf==2.3.0
open-spiel==1.4
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.8
pgx==2.0.0
pgx-minatar==0.5.0
protobuf==4.25.1
psutil==5.9.7
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
scipy==1.11.4
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
svgwrite==1.4.3
tabulate==0.9.0
tensorflow-probability==0.23.0
tensorstore==0.1.52
toolz==0.12.0
typing_extensions==4.9.0
urllib3==2.1.0
wandb==0.16.1
zipp==3.17.0
langchain
langchain_openai
python_dotenv